Add escape clause for slack usernames
During a live online demo, I tried to use the `tweet` skill to tweet my own twitter username @haacked. Well, that happens to match my Slack username, so it replaced it with `<@U....>` (my slack mention).

Well that's no good!

Instead, we want to detect that and block it. We provide an escape hatch. Just use `@@haacked` instead.