Use `bot.room.cache_key` for the cache key.
The `cache_key` property attempts to retrieve the room id. The room id is more stable for a cache key as it won't change even if you rename the room.

However some platforms (looking at you MS Teams) don't seem to give us a room id. So `cache_key` will use the `room_name` as a backup.

This is a breaking change as it will require users of this skill to re-authenticate.